Testing multiple robot servers
They do not work well together since the logic is to go to the closest
customer. Need the centralize controller to fix this issue.
Had to change return from customer to make sure the robot knew an order
was already taken.
There was a crash in the kitchen area with a strange id value of
negative some big number. I have not tracked down the issue yet.